Updated VAMOS identification

+++ Lib/src/Analysis.cxx
@@ -33,10 +33,14 @@ Analysis::~Analysis()
 bool Analysis::RunAnalysis()
 {
     //Creating needed files
-    system("cd Configs/Cuts/; root -l -b -q E_TOF_cuts.cxx; cd -;");
-    system("cd Configs/Cuts/; root -l -b -q DE_E_cuts.cxx; cd -;");
-    system("cd Configs/Cuts/; root -l -b -q MQ_Q_cuts.cxx; cd -;");
-    //system("cd Configs/Interpolations/; root -l -b -q TW_Brho_M46_Z18.C; cd -;");
+    int status;
+    status = system("cd Configs/Cuts/ && root -l -b -q E_TOF_cuts.cxx; cd -;");
+    if (!(WIFEXITED(status))) throw std::runtime_error("unable to run command\n");
+    status = system("cd Configs/Cuts/ && root -l -b -q DE_E_cuts.cxx; cd -;");
+    if (!(WIFEXITED(status))) throw std::runtime_error("unable to run command\n");
+    status = system("cd Configs/Cuts/ && root -l -b -q MQ_Q_cuts.cxx; cd -;");
+    if (!(WIFEXITED(status))) throw std::runtime_error("unable to run command\n");
+    //system("cd Configs/Interpolations/ && root -l -b -q TW_Brho_M46_Z18.C; cd -;");
 
     //Removing previous data
     system("rm -rf ./Out/*");
